requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
oauthlib==3.1.0
py-trello==0.17.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.2
